% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatchTable.R
\name{CatchTable}
\alias{CatchTable}
\title{Extract the Catch and effort data as a flattend data.frame}
\usage{
CatchTable(logdata, SelBank, Area = NA, tableNames = FALSE)
}
\arguments{
\item{logdata}{is the processed log data from the functions GetLogData() and ProcessLogData().}

\item{SelBank}{a numeric delimiter of the assessment region (1 = Banquereau and 2 = Grand Bank).}

\item{Area}{(default=NA) is a optional indicator to flag the size of the assessment area (bank). If NA Banquereau and Grand Bank are assumed to be 10908.1 & 49473.0 km2 respectively}

\item{tableNames}{defines whether the table names should be returned as expression (e.g. superscripts)}
}
\description{
CatchTable
}
\author{
Brad Hubley and Ryan Stanley
}

