% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CpuePlot.R
\name{CpuePlot}
\alias{CpuePlot}
\title{Plot the annual CPUE and associated trigger level.}
\usage{
CpuePlot(logdata, SelBank, stat = "mean", returnData = FALSE,
  anon = FALSE)
}
\arguments{
\item{logdata}{is the processed log data from the functions GetLogData() and ProcessLogData().}

\item{SelBank}{a numeric delimiter of the assessment region (1 = Banquereau and 2 = Grand Bank).}

\item{stat}{variable (default="mean") defining whether the plot returned should have all the ship specific means (stat="all")
in a given year or should represented as the average for a ship in a given year.
Note if variable anon = TRUE only the grouped means will be returned.}

\item{returnData}{(default=FALSE) if true the function will not return a plot but will instead return a
dataframe summarizing the yearly averaged landings.}

\item{anon}{(default=FALSE) should the output be anonymized for public consumption (no ship specific data).}
}
\description{
Function to subset loci and populations
}

