% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CumulativeFootprint.R
\name{cumulative_footprint}
\alias{cumulative_footprint}
\title{Plot the cumulative footprint at specified levels.}
\usage{
cumulative_footprint(CatchData, SelBank, stat = "footprint", Banksize = NA,
  lagtime = 5, returnData = FALSE)
}
\arguments{
\item{CatchData}{is average catch per year. Columns are ordered as "Year" "se" "sd" and "CPUE". CPUE is the mean CPUE in a givne year.
Note the required input can be generated by the CatchTable function}

\item{SelBank}{a numeric delimiter of the assessment region (1 = Banquereau and 2 = Grand Bank).}

\item{stat}{variable defining whether the plot will have the percent of the bank covered ('percent') or the spatial footprint ('footprint' : DEFAULT)}

\item{Banksize}{Coverge in square km of the bank curated by the Catchdata- catchTable() function.}

\item{lagtime}{is an integer or vector of integers specifying the desired lag-times in years (default = 5 years)}
}
\description{
Function will accumulate the footprint at specified time lags
}

