% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotLengthFreq.R
\name{PlotLengthFreq}
\alias{PlotLengthFreq}
\title{PPlot the annual length frequency distributions.}
\usage{
PlotLengthFreq(lfdata, SelBank, stat = "cdf")
}
\arguments{
\item{lfdata}{is the length frequency data from the database. Columns should include AREA (3 for Grand Bank and 4 for Banquereau)
YEAR, SHELL_LEN (length in mm) and NUM (frequency of this measurement in a given year).}

\item{SelBank}{a numeric delimiter of the assessment region (1 = Banquereau and 2 = Grand Bank).}

\item{stat}{variable which defines which plot to return. "cdf" returns the cumulative length frequency distribution
and "lf" returns the length frequency distribution.}
}
\description{
Function to subset loci and populations
}

