% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotPercentLarge.R
\name{PlotPercentLarge}
\alias{PlotPercentLarge}
\title{Plot the number and percent of large clams in a sample per year with associated trigger levels.}
\usage{
PlotPercentLarge(lfdata, SelBank, buffer = 3, plot_loess = FALSE)
}
\arguments{
\item{lfdata}{is the length frequency data from the database. Columns should include AREA (3 for Grand Bank and 4 for Banquereau)
YEAR, SHELL_LEN (length in mm) and NUM (frequency of this measurement in a given year).}

\item{SelBank}{a numeric delimiter of the assessment region (1 = Banquereau and 2 = Grand Bank).}

\item{buffer}{an integer which defines the number of percentage points above the maximum percent of large clams
in a given year to be added to the plot limits.}

\item{plot_loess}{Logical query (default = FALSE) specifying whether a loess smoothed regression should be added to the plot.}
}
\description{
Function to subset loci and populations
}

