% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMSfished.R
\name{OMSfished}
\alias{OMSfished}
\title{Plot the fishing effort as a function of one minute squared 'cells' fished annually.}
\usage{
OMSfished(logdata, SelBank)
}
\arguments{
\item{logdata}{is the processed log data from the functions GetLogData() and ProcessLogData().}

\item{SelBank}{a numeric delimiter of the assessment region (1 = Banquereau and 2 = Grand Bank).}
}
\description{
Function to subset loci and populations
}

